requirements.txt
pillow==10.3.0
pynput==1.7.6